requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
bitbucket-api==0.5.0
click==6.6
-e git+git@github.com:fprieur/vdmcli.git@4c73efe5836e28f3778fe270e10b970ee0c74e82#egg=click_example_repo
decorator==4.0.10
enum34==1.1.6
funcsigs==1.0.2
future==0.15.2
ipython==5.1.0
ipython-genutils==0.1.0
minitest==2.0.3
mock==2.0.0
oauthlib==1.1.2
pathlib2==2.1.0
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
ptyprocess==0.5.1
Pygments==2.1.3
pywatch==0.4
requests==2.11.1
requests-oauthlib==0.7.0
ruamel.ordereddict==0.4.9
ruamel.yaml==0.12.14
sh==1.11
simplegeneric==0.8.1
simplejson==3.9.0
six==1.10.0
traitlets==4.3.0
typing==3.5.2.2
uritemplate==0.6
-e git+git@github.com:fprieur/vdmcli.git@4c73efe5836e28f3778fe270e10b970ee0c74e82#egg=vdmcli
wcwidth==0.1.7
